[STEP 2] refactor: 각 도메인 패키지 추가 및 파일 이관
